git-svn-id: https://svn.localdev.cc/svn/shopgate_plugin_library/branches/input_fields@519 e652e40c-185c-4641-87fb-3ecfe257c510

+++ classes/core.php
@@ -3,7 +3,7 @@
 ###################################################################################
 # define constants
 ###################################################################################
-define('SHOPGATE_LIBRARY_VERSION', '2.0.29');
+define('SHOPGATE_LIBRARY_VERSION', '2.0.31');
 define('SHOPGATE_LIBRARY_ENCODING' , 'UTF-8');
 define('SHOPGATE_BASE_DIR', realpath(dirname(__FILE__).'/../'));
 define('SHOPGATE_ITUNES_URL', 'http://itunes.apple.com/de/app/shopgate-eine-app-alle-shops/id365287459?mt=8');
@@ -806,7 +806,7 @@ abstract class ShopgateObject {
 	 */
 	public function jsonEncode($value) {
 		// if json_encode exists use that
-		if (function_exists("json_encode")) {
+		if ( extension_loaded("json") && function_exists("json_decode") ) {
 			return $string = json_encode($value);
 		}
 
@@ -830,7 +830,7 @@ abstract class ShopgateObject {
 	 */
 	public function jsonDecode($json, $assoc = false) {
 		// if json_decode exists use that
-		if (function_exists("json_decode")) {
+		if ( extension_loaded("json") && function_exists("json_decode") ) {
 			return json_decode($json, $assoc);
 		}
 
@@ -1196,6 +1196,32 @@ class ShopgatePluginApi extends ShopgateObject {
 		return !(isset($this->response["error"]) && $this->response["error"] > 0);
 	}
 
+	/**
+	 * Open the given file and output line by line
+	 * to the default output
+	 *
+	 * @param string $fileName
+	 * @throws ShopgateLibraryException
+	 */
+	private function outputFile($fileName) {
+		// No output on streaming
+		if(preg_match("/^php/", $fileName)) return;
+		
+		if (!file_exists($fileName)) {
+			throw new ShopgateLibraryException(ShopgateLibraryException::PLUGIN_FILE_NOT_FOUND, 'File: '.$fileName);
+		}
+		
+		$fp = @fopen($fileName, "r");
+		if (!$fp) {
+			throw new ShopgateLibraryException(ShopgateLibraryException::PLUGIN_FILE_OPEN_ERROR, 'File: '.$fileName);
+		}
+		
+		// output csv file
+		while ($line = fgets($fp)) echo $line;
+		
+		// clean up and leave
+		fclose($fp);
+	}
 
 	/******************************************************************
 	 * Following methods represent the Shopgate Plugin API's actions: *
@@ -1444,6 +1470,11 @@ class ShopgatePluginApi extends ShopgateObject {
 	 * @see http://wiki.shopgate.com/Shopgate_Plugin_API_get_items_csv/de
 	 */
 	private function getItemsCsv() {
+		// output headers ...
+		header("HTTP/1.0 200 OK");
+		header('Content-Type: text/csv');
+		header('Content-Disposition: attachment; filename="items.csv"');
+		
 		$generate_csv = ($this->config["generate_items_csv_on_the_fly"] || isset($this->params["generate_items_csv_on_the_fly"]));
 
 		if (isset($this->params["limit"]) && isset($this->params["offset"])) {
@@ -1456,27 +1487,10 @@ class ShopgatePluginApi extends ShopgateObject {
 		if ($generate_csv) {
 			$this->plugin->startGetItemsCsv();
 		}
-
+		
 		$fileName = ShopgateConfig::getItemsCsvFilePath();
-		if (!file_exists($fileName)) {
-			throw new ShopgateLibraryException(ShopgateLibraryException::PLUGIN_FILE_NOT_FOUND, 'File: '.$fileName);
-		}
-
-		$fp = @fopen($fileName, "r");
-		if (!$fp) {
-			throw new ShopgateLibraryException(ShopgateLibraryException::PLUGIN_FILE_OPEN_ERROR, 'File: '.$fileName);
-		}
-
-		// output headers ...
-		header("HTTP/1.0 200 OK");
-		header('Content-Type: text/csv');
-		header('Content-Disposition: attachment; filename="items.csv"');
-
-		// ... and csv file
-		while ($line = fgets($fp)) echo $line;
-
-		// clean up and leave
-		fclose($fp);
+		$this->outputFile($fileName);
+		
 		exit;
 	}
 
@@ -1487,29 +1501,17 @@ class ShopgatePluginApi extends ShopgateObject {
 	 * @see http://wiki.shopgate.com/Shopgate_Plugin_API_get_categories_csv/de
 	 */
 	private function getCategoriesCsv() {
+		// output headers
+		header("HTTP/1.0 200 OK");
+		header('Content-Type: text/csv');
+		header('Content-Disposition: attachment; filename="categories.csv"');
+		
 		// generate / update categories csv file
 		$this->plugin->startGetCategoriesCsv();
 
 		$fileName = ShopgateConfig::getCategoriesCsvFilePath();
-		if (!file_exists($fileName)) {
-			throw new ShopgateLibraryException(ShopgateLibraryException::PLUGIN_FILE_NOT_FOUND, 'File: '.$fileName);
-		}
-
-		$fp = @fopen($fileName, "r");
-		if (!$fp) {
-			throw new ShopgateLibraryException(ShopgateLibraryException::PLUGIN_FILE_OPEN_ERROR, 'File: '.$fileName);
-		}
-
-		// output headers ...
-		header("HTTP/1.0 200 OK");
-		header('Content-Type: text/csv');
-		header('Content-Disposition: attachment; filename="categories.csv"');
-
-		// ... and csv file
-		while ($line = fgets($fp)) echo $line;
-
-		// clean up and leave
-		fclose($fp);
+		$this->outputFile($fileName);
+		
 		exit;
 	}
 
@@ -1520,29 +1522,17 @@ class ShopgatePluginApi extends ShopgateObject {
 	 * @see http://wiki.shopgate.com/Shopgate_Plugin_API_get_reviews_csv/de
 	 */
 	private function getReviewsCsv() {
+		// output headers
+		header("HTTP/1.0 200 OK");
+		header('Content-Type: text/csv');
+		header('Content-Disposition: attachment; filename="reviews.csv"');
+		
 		// generate / update reviews csv file
 		$this->plugin->startGetReviewsCsv();
 
 		$fileName = ShopgateConfig::getReviewsCsvFilePath();
-		if (!file_exists($fileName)) {
-			throw new ShopgateLibraryException(ShopgateLibraryException::PLUGIN_FILE_NOT_FOUND, 'File: '.$fileName);
-		}
-
-		$fp = @fopen($fileName, "r");
-		if (!$fp) {
-			throw new ShopgateLibraryException(ShopgateLibraryException::PLUGIN_FILE_OPEN_ERROR, 'File: '.$fileName);
-		}
-
-		// output headers ...
-		header("HTTP/1.0 200 OK");
-		header('Content-Type: text/csv');
-		header('Content-Disposition: attachment; filename="reviews.csv"');
-
-		// ... and csv file
-		while ($line = fgets($fp)) echo $line;
-
-		// clean up and leave
-		fclose($fp);
+		$this->outputFile($fileName);
+		
 		exit;
 	}
 
@@ -1553,26 +1543,16 @@ class ShopgatePluginApi extends ShopgateObject {
 	 * @see http://wiki.shopgate.com/Shopgate_Plugin_API_get_pages_csv/de
 	 */
 	private function getPagesCsv() {
-		$fileName = ShopgateConfig::getPagesCsvFilePath();
-		if (!file_exists($fileName)) {
-			throw new ShopgateLibraryException(ShopgateLibraryException::PLUGIN_FILE_NOT_FOUND, 'File: '.$fileName);
-		}
-
-		$fp = @fopen($fileName, "r");
-		if (!$fp) {
-			throw new ShopgateLibraryException(ShopgateLibraryException::PLUGIN_FILE_OPEN_ERROR, 'File: '.$fileName);
-		}
-
-		// output headers ...
+		// output headers
 		header("HTTP/1.0 200 OK");
 		header('Content-Type: text/csv');
 		header('Content-Disposition: attachment; filename="pages.csv"');
+		
+		$this->plugin->startGetPagesCsv();
+		
+		$fileName = ShopgateConfig::getPagesCsvFilePath();
+		$this->outputFile($fileName);
 
-		// ... and csv file
-		while ($line = fgets($fp)) echo $line;
-
-		// clean up and leave
-		fclose($fp);
 		exit;
 	}
 
@@ -2233,6 +2213,11 @@ abstract class ShopgatePlugin extends ShopgateObject {
 	/**
 	 * @var int
 	 */
+	private $rowCount = 0;
+	
+	/**
+	 * @var int
+	 */
 	private $bufferCounter = 0;
 
 	/**
@@ -2319,10 +2304,16 @@ abstract class ShopgatePlugin extends ShopgateObject {
 	 */
 	private final function createBuffer($filePath) {
 		$this->timeStart = time();
-		$filePath .= ".tmp";
+		$baseFile = $filePath;
+		
+		if(!preg_match("/^php/", $filePath)) {
+			$filePath .= ".tmp";
+			$baseFile = basename($baseFile);
+		}
 
-		$this->log('Trying to create "'.basename($filePath).'". ', 'access');
-
+		$this->log("Trying to create {$baseFile}", 'access');
+		
+		$this->rowCount = 0;
 		$this->fileHandle = @fopen($filePath, 'w');
 		if (!$this->fileHandle) {
 			throw new ShopgateLibraryException(ShopgateLibraryException::PLUGIN_FILE_OPEN_ERROR, 'File: '.$filePath);
@@ -2340,10 +2331,12 @@ abstract class ShopgatePlugin extends ShopgateObject {
 		$this->flushBuffer();
 		fclose($this->fileHandle);
 
-		if(file_exists($filePath)){
-			unlink($filePath);
+		if(!preg_match("/^php/", $filePath)) {
+			if(file_exists($filePath)){
+				unlink($filePath);
+			}
+			rename($filePath.".tmp", $filePath);
 		}
-		rename($filePath.".tmp", $filePath);
 
 		$this->log('Fertig, '.basename($filePath).' wurde erfolgreich erstellt', "access");
 		$duration = time() - $this->timeStart;
@@ -2394,12 +2387,14 @@ abstract class ShopgatePlugin extends ShopgateObject {
 		$this->finishBuffer(ShopgateConfig::getReviewsCsvFilePath());
 	}
 
+
+	
 	/**
 	 * Adds a line to the csv file buffer.
 	 *
 	 * @param mixed[] $itemArr
 	 */
-	protected final function addItem($itemArr) {
+	private final function addRow($itemArr) {
 		$this->buffer[] = $itemArr;
 		$this->bufferCounter++;
 
@@ -2411,6 +2406,41 @@ abstract class ShopgatePlugin extends ShopgateObject {
 			$this->flushBuffer();
 		}
 	}
+	
+	/**
+	 * @deprecated
+	 */
+	protected final function addItem($item) {
+		$this->addRow($item);
+	}
+	
+	/**
+	 * @see ShopgatePlugin::addRow
+	 * @param array $itemArr
+	 */
+	protected final function addItemRow($item) {
+		$item = array_merge( $this->buildDefaultItemRow(), $item );
+		
+		$this->addRow( $item );
+	}
+	/**
+	 * @see ShopgatePlugin::addRow
+	 * @param array $itemArr
+	 */
+	protected final function addCategoryRow($category) {
+		$category = array_merge($this->buildDefaultCategoryRow(), $category);
+		
+		$this->addRow($category);
+	}
+	/**
+	 * @see ShopgatePlugin::addRow
+	 * @param array $itemArr
+	 */
+	protected final function addReviewRow($review) {
+		$review = array_merge($this->buildDefaultReviewRow(), $review);
+		
+		$this->addRow($review);
+	}
 
 	/**
 	 * Flushes buffer to the currently opened file handle in $this->fileHandle.
@@ -2418,12 +2448,12 @@ abstract class ShopgatePlugin extends ShopgateObject {
 	 * The data is converted to utf-8 if mb_convert_encoding() exists
 	 */
 	private final function flushBuffer() {
-		if (empty($this->buffer) && ftell($this->fileHandle) == 0) {
+		if (empty($this->buffer) && $this->rowCount == 0) {
 			throw new ShopgateLibraryException(ShopgateLibraryException::PLUGIN_FILE_EMPTY_BUFFER);
 		}
 
 		// write headline if it's the beginning of the file
-		if (ftell($this->fileHandle) == 0) {
+		if ($this->rowCount == 0) {
 			fputcsv($this->fileHandle, array_keys($this->buffer[0]), ';', '"');
 		}
 
@@ -2433,6 +2463,7 @@ abstract class ShopgatePlugin extends ShopgateObject {
 			}
 
 			fputcsv($this->fileHandle, $item, ";", "\"");
+			$this->rowCount++;
 		}
 
 		$this->buffer = array();
@@ -2458,10 +2489,18 @@ abstract class ShopgatePlugin extends ShopgateObject {
 	}
 
 	/**
+	 * @see buildDefaultItemRow
+	 * @deprecated
+	 */
+	protected function buildDefaultProductRow() {
+		return $this->buildDefaultItemRow();
+	}
+	
+	/**
 	 * @return string[] An array with the csv file field names as indices and empty strings as values.
 	 * @see http://wiki.shopgate.com/CSV_File_Items/de
 	 */
-	protected function buildDefaultProductRow() {
+	protected function buildDefaultItemRow() {
 		$row = array(
 			/* responsible fields */
 			'item_number' 				=> "",
@@ -2490,6 +2529,7 @@ abstract class ShopgatePlugin extends ShopgateObject {
 			'use_stock' 				=> "0",
 			'stock_quantity' 			=> "",
 			'ean' 						=> "",
+			'isbn' 						=> "",
 			'pzn'						=> "",
 			'last_update' 				=> "",
 			'tags' 						=> "",
@@ -2608,7 +2648,7 @@ abstract class ShopgatePlugin extends ShopgateObject {
 	 * @return string[] An array with the csv file field names as indices and empty strings as values.
 	 * @see http://wiki.shopgate.com/CSV_File_Reviews/de
 	 */
-	protected function buildDefaultReviewsRow() {
+	protected function buildDefaultReviewRow() {
 		$row = array(
 			"item_number" => '',
 			"update_review_id" => '',
@@ -2620,6 +2660,14 @@ abstract class ShopgatePlugin extends ShopgateObject {
 		);
 
 		return $row;
+	}
+	
+	/**
+	 * @see buildDefaultReviewRow
+	 * @deprecated
+	 */
+	protected function buildDefaultReviewsRow() {
+		return $this->buildDefaultReviewRow();
 	}
 
 	/**
@@ -2685,6 +2733,78 @@ abstract class ShopgatePlugin extends ShopgateObject {
 
 	protected $exchangeRate = 1;
 
+	/**
+	 *
+	 * @param array $loaders
+	 * @param array $shopgateItemArray
+	 * @param mixed $dataObject or $dataArray to access
+	 */
+	protected final function executeLoaders(array $loaders)
+	{
+		$arguments = func_get_args();
+		array_shift($arguments);
+	
+		foreach ($loaders as $method) {
+			if (method_exists($this, $method)) {
+				$this->log("Call Function {$method}", self::LOGTYPE_DEBUG);
+				$arguments[0] = call_user_func_array( array( $this, $method ), $arguments );
+			}
+		}
+	
+		return $arguments[0];
+	}
+	
+	/**
+	 * Creates an array of corresponding helper method names, based on the export type given
+	 * @param string $exportType
+	 * @return array
+	 */
+	private final function getCreateCsvLoaders($subjectName) {
+		$actions = array();
+		$subjectName = trim($subjectName);
+		if(!empty($subjectName)) {
+			$subjectName = ($subjectName);
+			$methodName = 'buildDefault'.$this->camelize($subjectName, true).'Row';
+			if(method_exists($this, $methodName)) {
+				foreach(array_keys($this->{$methodName}() ) as $sKey) {
+					$actions[] = $subjectName."Export" . $this->camelize($sKey, true);
+				}
+			}
+		}
+	
+		return $actions;
+	}
+	
+	/**
+	 * Returns an array with the method names of all item-loaders
+	 *
+	 * Example: exportItemName, exportUnitAmount
+	 *
+	 * @return array
+	 */
+	protected final function getCreateItemsCsvLoaders() {
+		return $this->getCreateCsvLoaders("item");
+	}
+	
+	/**
+	 * Returns an array with the method names of all item-loaders
+	 *
+	 * Example: exportCategoryCategoryNumber, exportCategoryCategoryName
+	 *
+	 * @return array
+	 */
+	protected final function getCreateCategoriesCsvLoaders() {
+		return $this->getCreateCsvLoaders("category");
+	}
+	
+	/**
+	 * Returns an array with the method names of all item-loaders
+	 *
+	 * @return array
+	 */
+	protected final function getCreateReviewsCsvLoaders() {
+		return $this->getCreateCsvLoaders("review");
+	}
 
 	/*******************************************************************************
 	 * Following methods are the callbacks that need to be implemented by plugins. *
